+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/QueryPerfTest/MetricsAccumulator.cs
@@ -2,7 +2,6 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.Diagnostics;
     using System.Linq;
     using Microsoft.Azure.Cosmos;
     using Microsoft.Azure.Cosmos.Diagnostics;
