+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/QueryPerfTest/TransportStats.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.Azure.Cosmos.Tests
+{
+    using System.Collections.Generic;
+
+    internal class TransportStats
+    {
+        public List<RequestTimeline> RequestTimeline { get; set; }
+    }
+}
